feat: improve hook when dealing with many files
- More efficient

The previous logic read all the content of `.remarkignore` in memory,
then loop over every lines. The new logic reads and stops when a match
has been found, making it less expensive.

- Improved

The script works better with multiple files.

Before, the warning message was displayed for each file that was not
present in the `.remarkignore` file.

Now, whatever the number of files not present in `.remarkignore`, the
script will display only one warning with the list of files that must be
added to `.remarkignore`.